requirements.txt
requests>=2.26.0
beautifulsoup4>=4.9.3
pandas>=1.3.0
lxml>=4.9.0
python-dotenv>=0.19.0